+++ sauron.js
@@ -64,6 +64,7 @@ module.exports = {
     "promise/valid-params": "warn",
     "promise/prefer-await-to-callbacks": "warn",
 
-    "unicorn/no-fn-reference-in-iterator": "warn"
+    "unicorn/no-fn-reference-in-iterator": "warn",
+    "unicorn/consistent-function-scoping": "error"
   }
 };